backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Simon3640 <simon.garcial@udea.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
loguru = "^0.6.0"
SQLAlchemy = "^1.4.41"
uvicorn = "^0.18.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2-binary = "^2.9.4"
email-validator = "^1.3.0"
business-rule-engine = "^0.2.0"
python-multipart = "^0.0.5"
fastapi = "^0.85.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
